feat: move table into other folder
