+++ src/pages/ValidatorManagement/ManagementTable/ManagementTable.css
